FireBall
Added a fireball attack
Features
* Destroys enemy when hits enemy
* Destroys itself when hits enemy
* If it will not hit an enemy It will destroy itself within 3 seconds
