--- Assets/Enemy.cs
@@ -1,38 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Enemy : MonoBehaviour
-{
-
-    [SerializeField]
-    private int damage = 5;
-    [SerializeField]
-    private float speed = 1.5f;
-
-    [SerializeField]
-    private EnemyData data;
-
-    private GameObject player;
-
-
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        player = GameObject.FindGameObjectWithTag("Player");
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        Chase();
-    }
-
-    private void Chase()
-    {
-        transform.position = Vector3.MoveTowards(transform.position, player.transform.position, speed * Time.deltaTime);
-
-
-    }
-}
