--- Assets/EnemyData.cs
@@ -1,14 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-[CreateAssetMenu(fileName = "Data", menuName = "ScriptableObject/Enemy", order = 1)]
-
-public class EnemyData : ScriptableObject
-
-{
-    public int hp;
-    public int damage;
-    public float speed;
-
-}
