+++ Assets/Materials/FireBall.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ef9b2f84d1e855d4ca8fda10173e950b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
