+++ Assets/Scripts/DestroyAfter.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class DestroyAfter : MonoBehaviour
+{
+    public float lifeDuration;
+    void Start()
+    {
+        Destroy(gameObject, lifeDuration);
+    }
+
+}
