+++ Assets/Scripts/DestroyAfter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ac3e8f57982a3c6479930b3c717deb9d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
