+++ Assets/Scripts/Enemy.cs
@@ -1,0 +1,47 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Enemy : MonoBehaviour
+{
+
+    [SerializeField]
+    private int damage = 5;
+    [SerializeField]
+    private float speed = 1.5f; 
+
+    [SerializeField]
+    private EnemyData data;
+
+    private GameObject player;
+    private GameObject enemy;
+
+
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        player = GameObject.FindGameObjectWithTag("Player");
+    }
+
+    private void OnTriggerEnter(Collider other)
+    {
+        if (other.tag == "FireBall")
+        {
+            Destroy(gameObject);
+        }
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        Chase();
+    }
+
+    private void Chase()
+    {
+        transform.position = Vector3.MoveTowards(transform.position, player.transform.position, speed * Time.deltaTime);
+    }
+
+    
+}
