+++ Assets/Scripts/FireBall.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class FireBall : MonoBehaviour
+{
+    private void OnTriggerEnter(Collider other)
+    {
+        if (other.tag == "Enemy")
+        {    
+            Destroy(gameObject);
+        }
+    }
+}
