+++ Assets/Scripts/FireBall.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1b4f6ae5f4e7270449df294458547aa2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
