+++ Assets/Scripts/LookAtPlayer.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class LookAtPlayer : MonoBehaviour
+{
+
+
+    public Transform Player;
+
+    void Update()
+    {
+        this.gameObject.transform.LookAt(Player);
+    }
+}
