+++ Assets/Scripts/PlayerAttacks.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e43c976d340769f43995b6d1817b092a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
